refactor: separate recovery request parsing from queue state
